feature: add a CLI implementation
cmd.js - Finds the Node.js executable in the current shell and constructs an interpreter session. By default, it works the same as running interactive.js; starts a Fever shell in interactive mode.

If a file path is given as an argument, that file will be read line-by-line and interpreted.

If the -i or --interactive flags are given, then interactive mode will start after evaluating the file argument.

Includes some basic help text.